+++ src/main/java/ajc/formation/spring/bibliotheque/repositories/AdministrateurRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ajc.formation.spring.bibliotheque.entities.Administrateur;
 
 
-public interface AdministrateurRepository extends JpaRepository<Administrateur, Long>{
+public interface AdministrateurRepository extends JpaRepository<Administrateur, Integer>{
 	Optional<Administrateur> findByLogin(String login);
 
 }
